+++ mlxtend/evaluate/__init__.py
@@ -7,4 +7,8 @@
 from .decision_regions import plot_decision_regions
 from .learning_curves import plot_learning_curves
 from .scoring import scoring
-__all__ = ["plot_decision_regions", "plot_learning_curves", "scoring"]
+from .confusion_matrix import confusion_matrix
+from .plot_confusion_matrix import plot_confusion_matrix
+
+__all__ = ["plot_decision_regions", "plot_learning_curves",
+           "scoring", "confusion_matrix", "plot_confusion_matrix"]